requirements.txt.txt
# requirements.txt
fastapi
uvicorn
psycopg[binary]
python-dotenv
google-generativeai